requirements.txt
# Core dependencies
ebooklib>=0.20
beautifulsoup4>=4.12.3
lxml>=6.0.2

# LlamaIndex core
llama-index-core>=0.11.0

# Vector store
chromadb>=0.5.0
llama-index-vector-stores-chroma>=0.3.0

# Ollama integration
llama-index-llms-ollama>=0.3.0
llama-index-embeddings-ollama>=0.3.0
